requirements.txt
# Core Slack and Databricks dependencies
slack-bolt>=1.18.0
slack-sdk>=3.20.0  
databricks-sdk>=0.15.0

# Web framework
flask>=2.0.0

# System monitoring for optimizations
psutil>=5.8.0

# Additional dependencies for enhanced functionality
python-dotenv>=0.19.0
requests>=2.25.0 